app/src/main/java/com/phpbg/easysync/dav/Resource.kt
/*
 * MIT License
 *
 * Copyright (c) 2023 Samuel CHEMLA
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package com.phpbg.easysync.dav

import java.time.Instant

data class Resource(
    val rootPath: RootPath,
    val href: String? = null,
    val creationdate: Instant? = null,
    val getlastmodified: Instant? = null,
    val isCollection: Boolean = false,
    val getetag: String? = null,
    val getcontentlength: String? = null,
    val getcontenttype: String? = null
) {
    val relativeHref: Path
        get() {
            if (this.href == null) {
                throw Exception()
            }
            val relativeHref = this.href.substringAfter(this.rootPath.path)
            return if (this.isCollection) CollectionPath(relativeHref) else File(relativeHref)
        }
}
